\name{fastPdistC}
\alias{fastPdistC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
fastPdist
}
\description{
calculates pairwise euclidean distances
}
\usage{
fastPdistC(Ar,Br)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Ar}{
[1:n,1:m]      data to calculate distances to
}
\item{Br}{
[1:n,1:m]      data to calculate distances to
}
}

\value{
dist[1:n,1:n]   distances
}
\references{
\url{https://blog.felixriedel.com/2013/05/pairwise-distances-in-r/}
}
\author{
Felix Riede
}
\keyword{fastPdist}

