\name{Hearingloss_N109}
\alias{Hearingloss_N109}
\docType{data}
\title{
Hearingloss data
}
\description{
Hearingloss data, with Gene2GoTerm matrix.
}
\usage{data('Hearingloss_N109')}

\details{
FeatureMarix_Gene2Term contains the dataset, NCBI are the row names for the genes and GoTerm_Header contains the column names for the GoTerms. Size of data matrix is 109 with dimension 829.
}
\source{
\href{https://www.ncbi.nlm.nih.gov/gtr/tests/509148/.Online}{NCBI OtoGenome Test for Hearing Loss}, accessed 24 June 2022.
}
\references{
GeneTestingRegistry (2018). OtoGenome Test for Hearing Loss Retrieved 2017
}
\examples{
data(Hearingloss_N109)
str(Hearingloss_N109)
}
\keyword{datasets}
\keyword{GoTerm}
\keyword{Genes}

